requirements.txt
Flask
flask-cors
requests
llama-cpp-python
huggingface-hub